4_2_HW3_optional

+++ src/main/java/net/nicken/repository/jdbc/JdbcMealRepositoryImpl.java
@@ -63,7 +63,7 @@ public class JdbcMealRepositoryImpl implements MealRepository{
 
     @Override
     public boolean delete(int id, int userId) {
-        return jdbcTemplate.update("DELETE FROM meals WHERE id=?", id, userId) !=0;
+        return jdbcTemplate.update("DELETE FROM meals WHERE id=? AND user_id=?", id, userId) != 0;
     }
 
     @Override
